Struct ord_subset::OrdVar

pub struct OrdVar<T: PartialOrd + PartialEq>(_);

Wrapper to signal that the contained variables have a total order. It's illegal to compare two OrdVars that are not ordered. For this reason, it's unsafe to create OrdVars without checking. Checked constructors are available for OrdSubset types.

Panics

Panics when a.partial_cmp(b) returns None for two values a,b.

Methods

impl<T: PartialOrd + PartialEq> OrdVar<T>

fn new(data: T) -> OrdVar<T> where T: Debug + OrdSubset

Construct an OrdVar out of the argument.

Panics

Panics if the argument is outside of the total order.

fn new_checked(data: T) -> Option<OrdVar<T>> where T: OrdSubset

Constructs an Option<OrdVar> out of the argument. Returns None if the argument is outside the total order.

unsafe fn new_unchecked(data: T) -> OrdVar<T>

Constructs an OrdVar without validity check. Incorrectly constructed OrdVars may panic during comparisons. This is the only way to construct an OrdVar out of a type that is not OrdSubset.

fn into_inner(self) -> T