Trait ord_subset::OrdSubset

pub trait OrdSubset: PartialOrd<Self> + PartialEq<Self> {
    fn is_outside_order(&self) -> bool;
}

Trait for types that form a total order when a few values are disallowed.

is_outside_order() must return true for these outliers and false for anything else.

std::cmp::PartialOrd::partial_cmp(a,b) must return Some(_) if a,b are both inside order and None if only one is outside order. Return value for two variables outside order is undefined.

Required Methods

fn is_outside_order(&self) -> bool